Split cycle import file helpers
