+++ js/cycle-import.js
@@ -23,6 +23,13 @@ import {
   parseNaturalCyclesCsvBundle,
 } from './cycle-import-adapters.js';
 import {
+  appleHealthArchiveEntry,
+  buildCycleFileContext,
+  clueArchiveEntries,
+  cycleFileKind,
+  naturalCyclesArchiveEntries,
+} from './cycle-import-file.js';
+import {
   clearCycleImport,
   clearCycleDB,
   clearCycleSource,
@@ -56,7 +63,6 @@ const appWindow = /** @type {Window & typeof globalThis & {
   navigate?: (category: string) => void,
   renderProfileButton?: () => void,
   closeModal?: () => void,
-  JSZip?: any,
 }} */ (typeof window !== 'undefined' ? window : {});
 
 let pendingCycleImport = null;
@@ -243,68 +249,8 @@ export async function parseAppleHealthCycleBlob(blob, fileName = 'apple-health-e
   return finalizeAppleHealthCycleImport(byKey, fileName);
 }
 
-let jszipLoad = null;
-function loadJSZip() {
-  if (appWindow.JSZip) return Promise.resolve(appWindow.JSZip);
-  if (jszipLoad) return jszipLoad;
-  jszipLoad = new Promise((resolve, reject) => {
-    const script = document.createElement('script');
-    script.src = '/vendor/jszip.min.js';
-    script.onload = () => appWindow.JSZip ? resolve(appWindow.JSZip) : reject(new Error('JSZip failed to load'));
-    script.onerror = () => reject(new Error('Failed to load /vendor/jszip.min.js'));
-    document.head.appendChild(script);
-  }).catch(err => {
-    jszipLoad = null;
-    throw err;
-  });
-  return jszipLoad;
-}
-
-function cycleFileKind(file) {
-  const name = String(file?.name || '').toLowerCase();
-  if (name.endsWith('.zip') || /zip/.test(file?.type || '')) return 'zip';
-  if (name.endsWith('.xml') || /xml/.test(file?.type || '')) return 'xml';
-  if (name.endsWith('.json') || name.endsWith('.cluedata') || file?.type === 'application/json') return 'json';
-  if (name.endsWith('.csv') || file?.type === 'text/csv') return 'csv';
-  return 'text';
-}
 export function isAppleHealthCycleFile(file) {
   return cycleFileKind(file) === 'xml';
-}
-
-async function buildCycleFileContext(file) {
-  const kind = cycleFileKind(file);
-  const context = { file, kind, text: null, archive: null, entries: [] };
-  if (kind === 'zip') {
-    const JSZip = await loadJSZip();
-    try {
-      context.archive = await JSZip.loadAsync(file);
-    } catch (err) {
-      if (/encrypt|password/i.test(String(err?.message || err))) {
-        throw new Error('This cycle ZIP is password-protected. Extract it first, then import the Clue JSON file inside.');
-      }
-      throw err;
-    }
-    context.entries = Object.values(context.archive.files || {}).filter(entry => !entry.dir);
-  } else if (kind !== 'xml') {
-    context.text = await file.text();
-  }
-  return context;
-}
-
-function appleHealthArchiveEntry(context) {
-  return context.entries.find(entry => {
-    const name = String(entry.name || '').toLowerCase();
-    return name === 'export.xml' || name === 'apple_health_export/export.xml';
-  }) || null;
-}
-
-function clueArchiveEntries(context) {
-  return context.entries.filter(entry => /(?:\.json|\.cluedata)$/i.test(entry.name || '') || /clue/i.test(entry.name || ''));
-}
-
-function naturalCyclesArchiveEntries(context) {
-  return context.entries.filter(entry => /\.csv$/i.test(entry.name || ''));
 }
 
 export const CYCLE_IMPORT_ADAPTERS = Object.freeze([
