+++ service-worker.js
@@ -134,6 +134,7 @@ const APP_SHELL = [
   '/js/supplements.js',
   '/js/supplement-impact.js',
   '/js/cycle-import-adapters.js',
+  '/js/cycle-import-file.js',
   '/js/cycle-import.js',
   '/js/cycle-store.js',
   '/js/cycle-summary.js',
